log(P tensor) for phi and chi
This plot shows log(chi P_tensor) and log(phi P_tensor) for several k values (the 0 mode is plotted in black). The calculation proceeds the same as in the plot phi_log_p_tensor.png, but the box size for phi_chi_log_p_tensor.png and at ICs was less than the Hubble length.