phi, chi, zeta trajectories
This plots some sampled trajectories of \phi, \chi, and \zeta at a sample of lattice sites in a trapped inflation potential (\Delta V_{trapped} = 1/2*g^2(\phi-\phi_p)^2). The horizontal axis is scaled so that \phi_p is located at ln(a/a_p) = 0. These plots are here mostly to compare to "trajectories_wdv83.png" which includes the instability.