phi, chi, and zeta trajectories
This plots \phi, \chi, and \zeta trajectories at sampled lattice sites (the same sites as used in "trajectories_wdv81_trapped.png"). The interaction term in the potential is of the form \Delta V = 1/2(g^2(\phi-\phi_p)^2 - b^2)\chi^2 (using b^2/g^2 = 1.e-3 M_p). The effects of the instability on \phi and \zeta are difficult to see on this plot: the slope of \phi shallows after the instability, the slopes of \zeta become overall more negative after the instability ( "trajectories_dif_wdv83_81.png" shows this more clearly").