Power in phi, chi, and zeta
Different coloured points correspond to different times after the instability. On the horizontal axis is k (comoving) in machine units. The peak value of P_{\phi} lags that of P_{\chi} in time (presumably this is due to rescattering). The centre of the P_{\chi} peak shifts to lower k after the instability. The effect on P_{\zeta} is small, but can be seen around k~60-110.